Check for nonzero exit status in tests.
If you specify :expect_failure => true to the 'ey' method, it will
allow ey to exit with a nonzero status code. Otherwise, a nonzero exit
will cause an exception to be raised.

Getting this change required us to pull in the open4 gem, which seems
to make mongrel segfault under ruby 1.8.6p287 on Mac OS X 10.5. This
only affects tests; the gem itself still works.

Change-Id: I797b28436e30254e5680ef3da9e01da5e2a293ab
Reviewed-on: http://review.engineyard.com/240
Reviewed-by: Andy Delcambre <adelcambre@engineyard.com>
Tested-by: Andy Delcambre <adelcambre@engineyard.com>