Allow multiple versions of ey-deploy.
Invoke the right one with "eysd _VERSION_ deploy ..."

Depend on a specific version of ey-deploy instead of a fuzzy match.

This avoids the upgrade problem; if two devs are deploying on an
instance and one upgrades to the latest engineyard, the other can keep
using their version. No more "server-side component is too new".

This is especially helpful when one of the two devs is cloud itself;
we can release a new version of ey-deploy and get customers using it
without breaking web deploys that use ey-deploy. Therefore, we don't
need coordinated releases of cloud and ey-deploy.

[#138]