Get it mostly running on Rubinius
FakeFS plays absolute havoc with #require and #autoload on Rubinius, so
a number of these are just to force those files into the system early to
work around FakeFS breaking things.

There are a few places where "ruby" is hardcoded, and I've used
Gem.ruby to properly pivot and use the current one.