Forced an addressable version
The engineyard-cloud-client gem requires that we list some stuff
in the Gemfile, including gitable.

The gitable gem requires addressable >= 2.2.7, which pretty much
causes bundler to ignore the fuzzy versions listed in the other
gems that we depend on, meaning that addressable-latest (2.4.0)
gets installed by bundler. Unfortunately, gitable doesn't work with
anything beyond addressable ~> 2.3

In the end, I've specified a hard addressable version and added
fuzzy versions to all dependencies in the gemspec